A version of lastIndexOf would crash the application if the character to be searched was not found in the string
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@175342 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/util/XMLString.cpp
@@ -1226,7 +1226,7 @@ int XMLString::lastIndexOf(const XMLCh ch,
                            const XMLCh* const toSearch,
                            const unsigned int toSearchLen)
 {
-    for (unsigned int i = toSearchLen-1; i >= 0; i--)
+    for (int i = (int)toSearchLen-1; i >= 0; i--)
     {
         if (toSearch[i] == ch)
             return i;
